pc-builder/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
APScheduler==3.10.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic[email]
python-multipart

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
pytest-cov==4.1.0
pytest-env==1.1.1